Fixed crash in border case.
If research that was not at the top of the queue happened to finish, the
queue may have elements in it but none of them is set as the next active
research. This results in a crash due to the incomplete refactorization of several functions.

This fix addresses the crash but not the lack of automatic researching.