Updated mod to use new research APIs.
LuaForce::current_research is a read-only property now.
I added a toggle for the native research queue to allow a smoother
experience. Of course, enabling the native research queue adds
complexity to the way techonologies are handled. As of this commit, the
mod's approach is to prioritize technologies in the native queue and
only kick in when the native queue is emptied.